+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [1.2.0] - 2018-10-03
+### Added
+* **changepw** action
+    * implements the AoratoPw user password reset from a TGT .kirbi
+    * equivalent to Kekeo's misc::changepw function
+
+
 ## [1.1.0] - 2018-09-31
 ### Added
 * **asktgs** action - takes /ptt:X, /dc:X, /ticket:X flags like asktgt, /service:X takes one or more SPN specifications